--- a/.babelrc
+++ b/.babelrc
@@ -1,1 +1,0 @@
-{ "presets": ["react", "es2015"] }