--- a/.gitignore
+++ b/.gitignore
@@ -7,3 +7,4 @@ Thumbs.db
 /releases/
 /tmp/
 docs/_site
+app/package.json