--- a/app/app.html
+++ b/app/app.html
@@ -3,9 +3,6 @@
   <head>
       <meta charset="utf-8">
       <title>Zazu</title>
-
-      <link href="stylesheets/main.css" rel="stylesheet" type="text/css">
-
       <script src="helpers/context_menu.js"></script>
       <script src="helpers/external_links.js"></script>
   </head>