--- a/app/app.js
+++ b/app/app.js
@@ -1,7 +1,7 @@
-import ReactDOM from 'react-dom'
-import React from 'react'
+const ReactDOM = require('react-dom')
+const React = require('react')
 
-import Zazu from './zazu'
+const Zazu = require('./zazu')
 
 ReactDOM.render(
   <Zazu />,