--- a/app/components/search.js
+++ b/app/components/search.js
@@ -21,14 +21,13 @@ const Search = React.createClass({
 
   render () {
     const { value } = this.props
-    return (
-      <input
-        className='mousetrap'
-        ref={this.setReference}
-        type='text'
-        onChange={this.handleQueryChange}
-        value={value} />
-    )
+    return React.createElement('input', {
+      className: 'mousetrap',
+      ref: this.setReference,
+      type: 'text',
+      onChange: this.handleQueryChange,
+      value: value,
+    })
   },
 })