--- a/docs/.spelling
+++ b/docs/.spelling
@@ -5,5 +5,6 @@
 # where filename is relative to this configuration file
 Zazu
 repo
+boolean
  - _documentation/blocks.md
 toc