--- a/package.json
+++ b/package.json
@@ -38,7 +38,7 @@
     "build": "gulp build",
     "release": "gulp release --env=production",
     "start": "gulp start",
-    "test": "electron-mocha --recursive --compilers js:babel-register test",
+    "test": "electron-mocha --recursive test",
     "lint": "eslint .",
     "install-native": "node ./tasks/install_native_module",
     "docs:publish": "gh-pages -t -d docs"