--- a/package.json
+++ b/package.json
@@ -16,7 +16,6 @@
     "gulp": "^3.9.0",
     "gulp-babel": "^6.1.2",
     "gulp-batch": "^1.0.5",
-    "gulp-less": "^3.0.3",
     "gulp-plumber": "^1.1.0",
     "gulp-util": "^3.0.6",
     "gulp-watch": "^4.3.5",