--- a/package.json
+++ b/package.json
@@ -1,8 +1,6 @@
 {
   "devDependencies": {
     "asar": "^0.10.0",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-preset-react": "^6.5.0",
     "chai": "^3.5.0",
     "electron-mocha": "^1.0.0",
     "electron-prebuilt": "^0.37.4",
@@ -14,11 +12,7 @@
     "fs-jetpack": "^0.7.0",
     "gh-pages": "^0.11.0",
     "gulp": "^3.9.0",
-    "gulp-babel": "^6.1.2",
-    "gulp-batch": "^1.0.5",
-    "gulp-plumber": "^1.1.0",
     "gulp-util": "^3.0.6",
-    "gulp-watch": "^4.3.5",
     "husky": "^0.11.4",
     "q": "^1.4.1",
     "rewire": "^2.5.1",
@@ -33,7 +27,7 @@
   },
   "scripts": {
     "prepush": "npm run lint && npm test",
-    "postinstall": "cd app && npm install",
+    "postinstall": "npm run build && cd app && npm install",
     "build": "gulp build",
     "release": "gulp release --env=production",
     "start": "gulp start",