--- a/tasks/release/osx.js
+++ b/tasks/release/osx.js
@@ -36,9 +36,9 @@ var cleanupRuntime = function () {
 var packageBuiltApp = function () {
   var deferred = Q.defer()
 
-  asar.createPackageWithOptions(projectDir.path('build'), finalAppDir.path('Contents/Resources/app.asar'), {
+  asar.createPackageWithOptions(projectDir.path('app'), finalAppDir.path('Contents/Resources/app.asar'), {
     dot: true,
-    unpack: projectDir.path('build/node_modules/node-notifier/vendor/**'),
+    unpack: projectDir.path('app/node_modules/node-notifier/vendor/**'),
   }, function () {
     deferred.resolve()
   })