--- a/tasks/start.js
+++ b/tasks/start.js
@@ -4,8 +4,8 @@ var childProcess = require('child_process')
 var electron = require('electron-prebuilt')
 var gulp = require('gulp')
 
-gulp.task('start', ['build', 'watch'], function () {
-  childProcess.spawn(electron, ['./build'], {
+gulp.task('start', ['build'], function () {
+  childProcess.spawn(electron, ['./app'], {
     stdio: 'inherit',
   })
   .on('close', function () {