--- a/test/lib/notification.spec.js
+++ b/test/lib/notification.spec.js
@@ -1,6 +1,6 @@
-import { expect } from 'chai'
-import notification from '../../app/lib/notification'
-import sinon from 'sinon'
+const { expect } = require('chai')
+const notification = require('../../app/lib/notification')
+const sinon = require('sinon')
 
 describe('Notification', () => {
   describe('tick', () => {