… app/package.tpl.json
… app/vendor/electron_boilerplate/dev_helper.js
… app/vendor/electron_boilerplate/window_state.js