|       | Before  | After |
| ----- | ------- | ----- |
| Test  | 1.38s   | 0.96s |
| Start | 3.48s   | 0.01s |